Move null folding logic to QL
